pyproject.toml
[project]
name = "madongdong"
version = "0.1.0"
description = "Personal blog system powered by FastAPI and Vue"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.49",
    "types-passlib>=1.7.7.20260211",
    "uvicorn[standard]>=0.44.0",
]
